NewestVersion HomeWork 10 .\lotto_3.py      55      5    91%
