+++ lotto_3.py
@@ -4,12 +4,12 @@ barrels = 90  #бочонков
 rows = 3      #строк в карточке
 cols = 5      #цифр в строке
 posits = 9    #всего позиций
-gamers = [['Вундеркинд', False], ['Железяка', True], ['Железяка 2', True]]
+gamers = [['Вундеркинд', False], ['Железяка', True]]
 
 class Card:
 
     def __init__(self, gamer):
-        self.positions = []     
+        self.positions = []
         self.numbers = random.sample(range(1, barrels+1), rows * cols)
         for i in range(rows):
             pos = random.sample(range(posits),cols)
@@ -66,31 +66,8 @@ class Bag:
     def next_barrel(self):
         self.cur_pos += 1
         if self.cur_pos >= barrels:
-            raise Exception('Бочонки RUN OUT - Кончились :)')
+            raise Exception('Мешок пуст')
         return self.bag[self.cur_pos], barrels-self.cur_pos-1
 
-def new_game():
-    games = []
-    for gamer in gamers:
-        games.append(Card(gamer))
-    print('>><<>>>  New game  >><<>>>')
-    loop = True
-    bag = Bag()
-    msg = ''
-    while loop:
-        barrel, rest_num = bag.next_barrel()
-        for game in games:
-            game.show()
-            des = game.decision(barrel, rest_num)
-            answer = game.cover(barrel, des)
-            loop = answer[0]
-            msg = answer[1]
-            if not loop:
-                break
-        if rest_num == 0:
-            loop = False
-            msg = 'GAME OVER!'
-    print(msg)
-new_game()
