+++ test_lotto3.py
@@ -1,6 +1,7 @@
 import pytest
 import unittest
 import random
+
 from lotto_3 import Card, Bag
 
 class TestCard(unittest.TestCase):
@@ -41,7 +42,7 @@ class TestBag(unittest.TestCase):
         self.assertNotEqual(self.bag.bag[0], self.bag.bag[-1])
         while self.bag.cur_pos < 89:
             self.assertNotEqual(self.bag.next_barrel(), self.bag.next_barrel())
-        with pytest.raises(Exception, match='Бочонки RUN OUT - Кончились :)'):
+        with pytest.raises(Exception, match='Мешок пуст'):
             self.bag.next_barrel()
         try:
             self.bag.next_barrel()
